test: integrate s3 tests in storage layer
